requirements.txt
Django==1.11.6
django-filter==1.1.0
django-widget-tweaks==1.4.1
olefile==0.44
Pillow==4.3.0
pytz==2017.3
